Fix Convex_hull_3 quickhull, with coplanar point
`CGAL::convex_hull_3` documents that the output is a triangulation
polyhedron. That is true... but for coplanar input points! In case of
coplanar input point, the output before this patch was a polyhedron with a
single polygonal face. The patch triangulates the face using a pivot point
and Euler operations.